reworked basically all pages to handle the 'errors' better and display the Debug (should be beautified; and fixed. It only shows the last 5 entries but the list doesnt limit itself, so SOMEBODY could spam that list with input simply by creating requests and thus crashing the page, i guess.), changed the css so the page doesn't look THAT  bad anymore.

+++ app/app.py
@@ -14,6 +14,7 @@ from main.constants import BASE_URL
 
 ##VARS
 app = Flask(__name__)
+debug_list = []
 
 ##METHODS
 
@@ -21,12 +22,12 @@ def get_readings_json():
     try:
         s = requests.get(BASE_URL + 'rest/reading/get/all')
         print(f"GET_READINGS - JSON: \n {s.json()}")
-        return s.json() 
+        return [s.json(), ["Fetched Readings.", s.status_code, s.headers.items()]]
     except requests.exceptions.ConnectionError as e:
         #logging.critical(e, exc_info=True) 
-        return None
+        return [None, ["Error fetching Readings.", "500", []]]
 
-def get_customers_json():
+def get_customer_json():
     """this just calls our rest-api and returns the json
         returns a list in the scheme of [json | None, (debug_message, debug_code, debug_headers)]
     """
@@ -36,7 +37,7 @@ def get_customers_json():
         return [s.json(), ["Fetched customers.", s.status_code, s.headers.items()]]
     except requests.exceptions.ConnectionError as e:
         #logging.critical(e, exc_info=True) 
-        return [None, ["Fetched customers.", "500", []]]
+        return [None, ["Error fetching customers customers.", "500", []]]
 
 
 #This doesnt have to be a route. @app.route('/customer/get/<id>/')
@@ -47,26 +48,26 @@ def get_customer_with_id(id):
         print(response.status_code)
 
         if (response.status_code == 200): #s.ok would theoretically work but 204 (no content) results in errors again that would need manual fixing.
-            return [response.json()[0], (response.content, response.status_code, response.headers.items())] #this only should hold one element so we can delte the []-brackets 
+            return [response.json()[0], [response.content, response.status_code, response.headers.items()]] #this only should hold one element so we can delte the []-brackets 
         else:
-            return [response.json()[0], (str(response.content), response.status_code, response.headers.items())]
-    except requests.exceptions.ConnectionError as e:
-        logging.critical(e, exc_info=True) 
-        return [None, ("Fetched customer via ID.", "500", [])]
-    #return s.json()   
-
-def get_users_json():
-    try: 
-        s = requests.get(BASE_URL + 'rest/users/get/all')
-        print(f"GET_USERS - JSON: \n {s.json()}")
-        return s.json() 
+            return [response.json()[0], [str(response.content), response.status_code, response.headers.items()]]
     except requests.exceptions.ConnectionError as e:
         #logging.critical(e, exc_info=True) 
-        return None
+        return [None,  ["Fetched customer via ID.", "500", []]]
+    #return s.json()   
+
+def get_user_json():
+    try: 
+        s = requests.get(BASE_URL + 'rest/user/get/all')
+        print(f"GET_USERS - JSON: \n {s.json()}")
+        return [s.json(), ["Fetched users.", s.status_code, s.headers.items()]]
+    except requests.exceptions.ConnectionError as e:
+        #logging.critical(e, exc_info=True) 
+        return [None, ["Error fetching customers customers.", "500", []]]
 
 def delete(o: Entity):
     try:
-        response = requests.get(f"{BASE_URL}rest/{o.get_db_type()}/delete/{o.id}")
+        response = requests.delete(f"{BASE_URL}rest/{o.get_db_type()}/delete/{o.id}")
         return [o, [response.content, response.status_code, response.headers.items()]]
     except requests.exceptions.ConnectionError as e:
         return [o, ["The server is unreachable.", 503, []]]
@@ -97,115 +98,128 @@ def create(o: Entity):
 def index():
     return render_template("Home.html")
 
-@app.route("/readings")
-def route_readings():
+@app.route("/reading")
+def route_reading():
+    readings_list, debug_info = get_readings_json()
+    print("?????????????????????")
+    print(readings_list)
+    debug_list.append(debug_info)
+    print(debug_list)
+    return render_template("Readings.html", debug_list=debug_list, readings_list=readings_list)
 
-    return render_template("Readings.html", readings_list=get_readings_json())
-
-@app.route("/readings/create", methods=["POST"])
-def route_readings_create():
+@app.route("/reading/create", methods=["POST"])
+def route_reading_create():
     print("CREATE-REQ-FORM: " + str(request.form))
     #fake-customer for request:
-    c  = Customer(id=request.form["customer"])
+    c  = Customer(id=int(request.form["customer"]))
     #none is 'better' than -1 i gues butttt does it maaatter?
-    r = Readings(id=None, comment=request.form["comment"], customer=c, dateofreading=request.form["dateofreading"], kindofmeter=request.form["kindofmeter"], metercount=request.form["metercount"], meterid=request.form["meterid"])
+    r = Readings(id=None, comment=request.form["comment"], customer=c, dateofreading=request.form["dateofreading"], kindofmeter=request.form["kindofmeter"], metercount=int(request.form["metercount"]), meterid=request.form["meterid"])
     print(str(r))
-    create(r)
-    return render_template("Readings.html")
+    r, resp = create(r)
 
-@app.route("/readings/edit", methods=["POST"])
-def route_readings_edit():
+    debug_list.append(resp) #TODO: why am i even dragging this shit here if it's a global-scope var now???
+    return route_reading()
+
+@app.route("/reading/edit", methods=["POST"])
+def route_reading_edit():
     print("EDIT-REQ: " + str((request.form)))
     
-    c  = Customer(id=request.form["customer"])
-    r = Readings(id=request.form["id"], comment=request.form["comment"], customer=c, dateofreading=request.form["dateofreading"], kindofmeter=request.form["kindofmeter"], metercount=request.form["metercount"], meterid=request.form["meterid"])
-    print(str(r))
-    edit(r)
-    return render_template("Readings.html")
+    c  = Customer(id=int(request.form["customer"])) #those casts feel useless, but VScode wants them if you turn type-checking to basic.
+    r = Readings(id=int(request.form["id"]), comment=request.form["comment"], customer=c, dateofreading=int(request.form["dateofreading"]), kindofmeter=request.form["kindofmeter"], metercount=int(request.form["metercount"]), meterid=request.form["meterid"])
+    r, resp = edit(r)
+    debug_list.append(resp)
+    return route_reading()
 
 @app.route('/readings/delete/<id>')
 def route_reading_delete(id):
-    delete(Readings(id=id))
+    _, resp = delete(Readings(id=id))
+    debug_list.append(resp)
     #delete_customer(id)
-    return render_template("Readings.html") 
+    return route_reading()
 
-@app.route('/customers')
-def route_customer(*args):
+@app.route('/customer')
+def route_customer():
     """
     routes to the customer page. Optionally takes a list of debug-messages [] in the first optional argument.
     if there are args, the debug-message of the fetch-all gets appended to that; otherwise just the fetch-all message gets returned as debug-message.
     """
-    print("Args")
-    print(args)
-    if len(args) > 0:
-        print(args[0])
+    
     ###TODO: if this fails; redirect to error page instead of... well. running into an error.
     ###is now done by customers.html itself
-    cList, debug_info = get_customers_json()
+    cList, debug_info = get_customer_json()
     print("DEBUG-INFO: ")
     print(debug_info)
-    if len(args) > 0:
-        args[0].append(debug_info)
-        print(args[0])
+    debug_list.append(debug_info)
+    print(debug_list)
 
     #this one is really ugly. Header is never used, either.
     #could be simplified by adding the other way around (no double-check on len(args))
-    return render_template("Customers.html", customer_list=cList, debug_info=args[0] if len(args) > 0 else [debug_info])
+    return render_template("Customer.html", customer_list=cList, debug_list=debug_list)
 
 
     
-@app.route('/customers/create', methods=['POST'])
+@app.route('/customer/create', methods=['POST'])
 def route_customer_create():
     #this is the actual api-call. DELETE has to be post for some browser-related-reasons i guess. It COULD be a normal method/route with a variable but yeaaah. #TODO.
     print("CREATE-REQ-FORM: " + str(request.form))
     c = Customer(firstname=request.form["firstname"], lastname=request.form["lastname"], id=-1)
     c, resp_val = create(c)
     
-    return route_customer([resp_val])
+    debug_list.append(resp_val)
+
+    return route_customer()
  
 
-@app.route('/customers/edit/', methods=['POST'])
+@app.route('/customer/edit/', methods=['POST'])
 def route_customer_edit():
     print("EDIT-REQ: " + str(request.form))
 
-    c = Customer(firstname=request.form["firstname"], lastname=request.form["lastname"], id=request.form["id"])
-    c, resp_val = edit(c)
+    c = Customer(firstname=request.form["firstname"], lastname=request.form["lastname"], id=int(request.form["id"]))
+    c, resp = edit(c)
+    debug_list.append(resp)
 
-    return route_customer([resp_val])
+    return route_customer()
 
 
-@app.route('/customers/delete/<id>')
+@app.route('/customer/delete/<id>')
 def route_customer_delete(id):
     c, resp_val = delete(Customer(firstname="", lastname="", id=id))
     #delete_customer(id)
-    return route_customer([resp_val])
+    debug_list.append(resp_val)
 
-@app.route('/users')
-def route_users():
-    return render_template("Users.html", users_list = get_users_json())
+    return route_customer()
 
-@app.route('/users/create', methods=["POST"])
-def route_users_create():
+@app.route('/user')
+def route_user():
+    uList, resp = get_user_json()
+    debug_list.append(resp) #TODO: Again; why am I doing this? Fcking masochist.
+    return render_template("User.html", users_list = uList, debug_list=debug_list)
+
+@app.route('/user/create', methods=["POST"])
+def route_user_create():
     print("CREATE-REQ-FORM: " + str(request.form))
     #print("0:" + request.form[0])
     #print("type:" +  request.form["type"])
     u = User(firstname=request.form["firstname"], lastname=request.form["lastname"], password=request.form["password"], token=request.form["token"], id=-1)
-    create(u)
-    return render_template("Users.html")
+    u, resp = create(u)
+    debug_list.append(resp)
+    return route_user()
 
-@app.route('/users/edit', methods=['POST'])
-def route_users_edit():
+@app.route('/user/edit', methods=['POST'])
+def route_user_edit():
     print("EDIT-REQ: " + str(request.form))
-    u = User(firstname=request.form["firstname"], lastname=request.form["lastname"], token=request.form["token"], password=request.form["password"], id=request.form["id"])    
-    edit(u)
+    u = User(firstname=request.form["firstname"], lastname=request.form["lastname"], token=request.form["token"], password=request.form["password"], id=int(request.form["id"]))    
+    u, resp = edit(u)
+    debug_list.append(resp)
+    return route_user()
 
-    return render_template("Users.html")
-
-@app.route('/users/delete/<id>')
+@app.route('/user/delete/<id>')
 def route_user_delete(id):
-    delete(User(firstname="", lastname="", id=id))
-    #delete_user(id)
-    return render_template("Users.html")
+    print(id)
+    u = User(firstname="", lastname="", id=id)
+    u, resp = delete(u)
+    debug_list.append(resp)
+    return route_user()
 
 
 @app.context_processor
@@ -217,18 +231,19 @@ def context_processors():
     @DeprecationWarning
     def print_readings():
         cList = get_readings_json()
-        str = """<table id='readings_table' class='object_table'>
-            <tr><th>COMMENT</th><th>CUSTOMER</th><th>kind</th><th>count</th><th>meter_id</th><th>sub</th><th>date</th><th>id</th><th>action</th></tr>"""
-        for r in cList:
-            str += f"""<tr><td>{r['comment']}</td><td>{r['customer']['id']}</td><td>{r['kindofmeter']}</td>
-            <td>{r['metercount']}</td><td>{r['meterid']}</td><td>{r['substitute']}</td><td>{r['dateofreading']}</td><td>{r['id']}</td>
-            <td><a href='/readings/delete/{r['id']}'>[delete]</a></td></tr>"""
-        str += "</table>"
-        return str
+        if (cList is not None):
+            str = """<table id='readings_table' class='object_table'>
+                <tr><th>COMMENT</th><th>CUSTOMER</th><th>kind</th><th>count</th><th>meter_id</th><th>sub</th><th>date</th><th>id</th><th>action</th></tr>"""
+            for r in cList:
+                str += f"""<tr><td>{r['comment']}</td><td>{r['customer']['id']}</td><td>{r['kindofmeter']}</td>
+                <td>{r['metercount']}</td><td>{r['meterid']}</td><td>{r['substitute']}</td><td>{r['dateofreading']}</td><td>{r['id']}</td>
+                <td><a href='/readings/delete/{r['id']}'>[delete]</a></td></tr>"""
+            str += "</table>"
+            return str
     
     @DeprecationWarning
     def print_customers():
-        cList = get_customers_json()
+        cList = get_customer_json()
         str = """<table id='customer_table' class='object_table'>
             <tr><th>LAST</th><th>FIRST</th><th>ID</th><th>action</th></tr>"""
         for c in cList:
@@ -238,7 +253,7 @@ def context_processors():
 
     @DeprecationWarning
     def print_users():
-        getUsers = get_users_json()
+        getUsers = get_user_json()
         if (getUsers[0] == 0): 
             return """<table id='user_table', class='object_table'>
             <tr><th>LAST</th><th>FIRST</th><th>ID</th></tr>
